+++ apps/web/src/app/(main)/interview-evaluation/timetable/[tab]/page.tsx
@@ -1,84 +1,47 @@
-'use client';
+// src/app/(main)/interview-evaluation/timetable/[tab]/page.tsx
+import React from 'react';
+import { notFound } from 'next/navigation';
+import { getServerSideTokens } from '@web/api/serverSideTokens';
+import { ServerFetchBoundary } from '@web/store/query/ServerFetchBoundary';
+import type { Tab } from '../../_components/CellRender/CellRenderer';
+import {
+  getMyTimeSlotsQueryOptions,
+  type MyTimeSlotsParams,
+} from '@web/store/query/useMyTimeSlotsQuery';
+import EvaluationTimetableClient from './EvaluationTimetableClient';
 
-import React, { useMemo } from 'react';
-import { useSearchParams, useRouter, useParams } from 'next/navigation';
-import { Flex, Text } from '@repo/ui';
-import { useMyTimeSlotsQuery } from '@web/store/query/useMyTimeSlotsQuery';
-import { TimeTable } from '@web/components/TimeTable/TimeTable';
-import { CellRenderer, Tab } from '../../_components/CellRender/CellRenderer';
-import DateNav from '@web/app/(main)/interview-management/_components/DateNav/DateNav';
+interface PageProps {
+  params: Promise<{ tab: string }>;
+  searchParams: Promise<{ interviewId?: string; date?: string }>;
+}
 
-export default function EvaluationTimetablePage() {
-  const router = useRouter();
-  const sp = useSearchParams();
-  const interviewId = Number(sp.get('interviewId') ?? '0');
-  const rawDate = sp.get('date');
-  const { tab } = useParams() as { tab: Tab };
+export default async function Page({ params, searchParams }: PageProps) {
+  const { tab: rawTab } = await params;
+  const tab = rawTab as Tab;
 
-  const activeDate = rawDate!.includes('-')
-    ? rawDate!.replace(/-/g, '.')
-    : rawDate;
+  const { interviewId: ivStr = '', date } = await searchParams;
 
-  const { data: schedules = [], isLoading } = useMyTimeSlotsQuery(interviewId);
-  if (isLoading) return null;
-
-  // 일치하는 스케줄 찾기
-  const schedule = schedules.find((s) => s.date === activeDate);
-  if (!schedule) {
-    return <Text>해당 날짜({activeDate})에 배정된 일정이 없습니다.</Text>;
+  const interviewId = ivStr ? Number(ivStr) : NaN;
+  if (isNaN(interviewId) || interviewId <= 0) {
+    return notFound();
   }
 
-  // 방별로 그룹핑
-  const { roomNames, timeSlots, startTime, endTime, interviewDuration } =
-    schedule;
-  const roomsMap = useMemo(() => {
-    return roomNames.reduce<Record<string, typeof timeSlots>>((acc, room) => {
-      acc[room] = timeSlots.filter((ts) => ts.roomName === room);
-      return acc;
-    }, {});
-  }, [roomNames, timeSlots]);
+  const { accessToken, refreshToken } = await getServerSideTokens();
+  const tokens = { accessToken, refreshToken };
 
-  // 날짜 이동
-  const handleDateChange = (nextDate: string) => {
-    const norm = nextDate.replace(/\./g, '-');
-    router.replace(
-      `/interview-evaluation/timetable/${tab}` +
-        `?interviewId=${interviewId}&date=${norm}`
-    );
-  };
+  //  SSR Prefetch: 내 타임슬롯 조회 옵션
+  const slotsOptions = getMyTimeSlotsQueryOptions({
+    interviewId,
+    tokens,
+  } as MyTimeSlotsParams);
 
   return (
-    <Flex direction="column" align="center" gap="4rem" width="100%">
-      <DateNav
-        dates={schedules.map((s) => s.date)}
-        active={activeDate!}
-        onChange={handleDateChange}
+    <ServerFetchBoundary fetchOptions={slotsOptions}>
+      <EvaluationTimetableClient
+        tab={tab}
+        interviewId={interviewId}
+        date={date}
       />
-
-      <Flex gap="4rem" justify="center" width="100%">
-        {roomNames.map((room) => (
-          <TimeTable
-            key={room}
-            title={room}
-            headers={tab === 'interviewer' ? ['지원자', '면접관'] : undefined}
-            startHour={Number(startTime.split(':')[0])}
-            endHour={Number(endTime.split(':')[0])}
-            interval={interviewDuration}
-            slots={roomsMap[room]!}
-            width={roomNames.length === 3 ? '31.3rem' : '40rem'}
-            renderCell={(row) => (
-              <CellRenderer
-                date={schedule.date}
-                row={row}
-                tab={tab}
-                slotData={roomsMap[room]!}
-                startHour={Number(startTime.split(':')[0])}
-                interval={interviewDuration}
-              />
-            )}
-          />
-        ))}
-      </Flex>
-    </Flex>
+    </ServerFetchBoundary>
   );
 }
