+++ apps/web/src/store/query/useMyTimeSlotsQuery.ts
@@ -1,35 +1,60 @@
 // src/store/query/useMyTimeSlotsQuery.ts
-import { useQuery, type UseQueryResult } from '@tanstack/react-query';
+import {
+  useQuery,
+  UseSuspenseQueryOptions,
+  queryOptions,
+  useSuspenseQuery,
+} from '@tanstack/react-query';
 import { GET } from '@web/api/fetch';
 import { queryKeys } from '../constants';
 import type { InterviewSchedule } from './useInterviewScheduleQuery';
+import { Tokens } from '@web/api/types';
+
+export interface MyTimeSlotsParams {
+  interviewId: number;
+  tokens?: Tokens;
+}
+
+const STALE_TIME = 1000 * 60 * 1;
+const GC_TIME = 1000 * 60 * 2;
 
 /**
- * 내 면접 배정 결과 조회 Query
- * → GET /api/v1/interviews/{interviewId}/my-time-slots
+ * @description
+ *   내 면접 배정 결과 조회용 React Query 옵션 생성
  */
-export function useMyTimeSlotsQuery(
-  interviewId: number
-): UseQueryResult<InterviewSchedule[], Error> {
-  return useQuery<InterviewSchedule[], Error>({
+export function getMyTimeSlotsQueryOptions({
+  interviewId,
+  tokens,
+}: MyTimeSlotsParams): UseSuspenseQueryOptions<InterviewSchedule[], Error> {
+  return queryOptions<InterviewSchedule[]>({
     queryKey: queryKeys.interview.myTimeSlots(interviewId),
-    queryFn: async () => {
-      const res = await GET<InterviewSchedule[]>(
-        `api/v1/interviews/${interviewId}/my-time-slots`
-      );
-      console.log('내 타임테이블', res.result);
-      return res.result.map((sch) => ({
-        ...sch,
-        // 날짜 포맷을 마침표(.)로 통일
-        date: sch.date.replace(/-/g, '.'),
-        timeSlots: sch.timeSlots.map((ts) => ({
-          ...ts,
-          // 고정 컬러
-          color: '#BECEFF',
-        })),
-      }));
-    },
-    staleTime: 1000 * 60, // 1분
+    queryFn: () =>
+      GET<InterviewSchedule[]>(
+        `api/v1/interviews/${interviewId}/my-time-slots`,
+        undefined,
+        tokens
+      ).then((res) =>
+        res.result.map((sch) => ({
+          ...sch,
+          // 날짜 포맷을 마침표(.)로 통일
+          date: sch.date.replace(/-/g, '.'),
+          timeSlots: sch.timeSlots.map((ts) => ({
+            ...ts,
+            // 고정 컬러
+            color: '#BECEFF',
+          })),
+        }))
+      ),
+    staleTime: STALE_TIME,
+    gcTime: GC_TIME,
     enabled: interviewId > 0,
   });
+}
+
+/**
+ * @description
+ *   Suspense 기반으로 내 면접 배정 결과를 조회하는 훅
+ */
+export function useMyTimeSlotsQuery(params: MyTimeSlotsParams) {
+  return useSuspenseQuery(getMyTimeSlotsQueryOptions(params));
 }
